Fixed migrations not rolling back if tables didn't have columns.

+++ src/db/migrations/20181226223221_connect_itemhistory_saldohistory.js
@@ -96,11 +96,13 @@ exports.up = async function(knex, Promise) {
     }
 };
 
-exports.down = function(knex, Promise) {
+exports.down = async function(knex, Promise) {
     if (process.env.NODE_ENV !== 'production') {
-        return knex.schema.table('ITEMHISTORY', (table) => {
-            table.dropColumn('saldhistid');
-        });
+        if (await knex.schema.hasColumn('ITEMHISTORY', 'saldhistid')) {
+            await knex.schema.table('ITEMHISTORY', (table) => {
+                table.dropColumn('saldhistid');
+            });
+        }
     } else {
         throw new Error('dont drop stuff in production');
     }
